feat: 增加事务和历史记录

+++ src/core/model/content.js
@@ -3,6 +3,28 @@ import { horizontalMove, verticalMove } from '../defaultActions/caretMove'
 import { del } from '../defaultActions/delete'
 import { input } from '../defaultActions/input'
 
+const mergeBlock = (o, n, shouldUpdates = []) => {
+  const oBlock = o.blockComponent
+  if (o.blockComponent !== n.blockComponent) {
+    if (n.len === 0) {
+      n.component.$editor.selection.rangePoints
+        .filter((point) => point.container === n)
+        .forEach((point) => {
+          if (point.pointName === 'start') {
+            point.range.setStart(n.nextLeaf, 0)
+          } else {
+            point.range.setEnd(n.nextLeaf, 0)
+          }
+        })
+    }
+    o.blockComponent.$path.insertChildrenAfter(n)
+    oBlock.$path.parent.component.update()
+    shouldUpdates.forEach((ins) => {
+      ins.component.update()
+    })
+  }
+}
+
 /**
  * @description 内容管理类
  * @export
@@ -10,13 +32,16 @@ import { input } from '../defaultActions/input'
  * @extends {Component}
  */
 export default class Content extends Component {
+  get renderContent () {
+    return this.$editor.formater.render(this.$path)
+  }
   /**
    * @description 类型
    * @readonly
    * @memberof Content
    * @instance
    */
-  get _type() {
+  get _type () {
     return 'inline'
   }
   constructor(props) {
@@ -29,7 +54,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  get contentLength() {
+  get contentLength () {
     return this.$path.len
   }
 
@@ -39,7 +64,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  get $path() {
+  get $path () {
     return this.props.path
   }
 
@@ -49,7 +74,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  get $editor() {
+  get $editor () {
     return this.props.editor
   }
 
@@ -58,7 +83,7 @@ export default class Content extends Component {
    * @memberof Content
    * @memberof Content
    */
-  onBeforeRender() {
+  onBeforeRender () {
     /*
      * 在diff的时候对于相同的组件类型的vdom,会被当成相同vnode，不会重新创建实例,只会切换props,
      * 但是一定会调用render，因此需要在这里切换path的组件上下文
@@ -73,7 +98,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  update(path, range) {
+  update (path, range) {
     // 执行更新前钩子
     this.onBeforeUpdate && this.onBeforeUpdate({ path: path || this.$path, range })
     return this.setState().then(() => {
@@ -91,7 +116,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  contentInput(path, range, data) {
+  contentInput (path, range, data) {
     const { offset, endContainer } = range
     path.insertData(offset, data)
     this._updatePoints(endContainer, offset, data.length)
@@ -102,38 +127,51 @@ export default class Content extends Component {
   }
 
   /**
-   * @description  删除动作
-   * @param {*} commonPath 最近公共路径
-   * @param {*} range 区间
-   * @memberof Content
-   * @instance
+   * @desc: 删除动作
+   * @param {*} commonPath
+   * @param {*} range
+   * @return {*}
    */
-  onContentDelete(commonPath, range) {
-    const { endContainer, endOffset, collapsed } = range
+  onContentDelete (commonPath, range) {
+    const { endContainer, endOffset, startContainer, startOffset, collapsed } = range
     // 选区折叠
     if (collapsed) {
       if (endOffset > 0) {
-        commonPath.onContentDelete(endOffset, 1)
-        if (commonPath.len === 0) {
-          this.onCaretLeave(commonPath, range, 'left')
-          commonPath.delete()
+        // 执行删除
+        startContainer.textDelete(endOffset, 1)
+        if (this.contentLength === 0) {
+          // 对于块级 当执行删除块内容为空时候 将被br填充 此时光标停留在段首
+          range.setStart(startContainer, 0)
+        } else if (startContainer.len === 0) {
+          const { path: prevSibling } = this.onCaretLeave(startContainer, range, 'left')
+          if (!prevSibling) return
+          if (prevSibling.blockComponent !== startContainer.blockComponent) {
+            range.setStart(startContainer, 0)
+          } else {
+            startContainer.delete()
+          }
         } else {
           this._updatePoints(endContainer, endOffset, -1)
         }
       } else {
-        const prevLeaf = commonPath.prevLeaf
+        const { path: prevSibling } = this.onCaretLeave(startContainer, range, 'left')
+        if (!prevSibling) return
         if (!this.contentLength) {
           const parent = this.$path.parent.component
           this.$path.delete()
           parent.update()
         }
-        if (prevLeaf) {
-          prevLeaf.component.onCaretEnter(prevLeaf, range, 'right')
-        }
+        mergeBlock(startContainer, prevSibling)
       }
+    } else if (startContainer === endContainer) {
+      startContainer.textDelete(endOffset, endOffset - startOffset)
     } else {
-      console.log('TODO')
+      startContainer.textDelete(startContainer.len, startContainer.len - startOffset)
+      endContainer.textDelete(endOffset, endOffset)
+      commonPath.deleteBetween(startContainer, endContainer)
+      mergeBlock(endContainer, startContainer)
     }
+    range.collapse(true)
     this.update(commonPath, range)
   }
 
@@ -146,7 +184,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretEnter(path, range, direction) {
+  onCaretEnter (path, range, direction) {
     if (direction === 'left') {
       let fromPath = path.prevLeaf
       if (!fromPath) return {}
@@ -167,12 +205,13 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretLeave(path, range, direction) {
+  onCaretLeave (path, range, direction) {
     if (direction === 'left') {
       let toPath = path.prevLeaf
       if (!toPath) return {}
       return toPath.component.onCaretEnter(toPath, range, 'right')
     } else {
+      console.log(this.$path.lastLeaf, path)
       // 从尾部离开
       let toPath = path.nextLeaf
       if (!toPath) return {}
@@ -187,7 +226,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretForward(path, range) {
+  onCaretForward (path, range) {
     range.offset < path.len && (range.offset += 1)
   }
 
@@ -199,7 +238,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretBackward(path, range) {
+  onCaretBackward (path, range) {
     range.offset > 0 && (range.offset -= 1)
   }
 
@@ -212,14 +251,14 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretMove(direction, range, event) {
+  onCaretMove (direction, range, event) {
     const path = range.container
     const caretMoveMethod = this[`${direction === 'left' ? 'onCaretBackward' : 'onCaretForward'}`]
     const { shiftKey } = event
     let res = { path, range }
     // 重置 d
     if (range.d === 0) range.d = direction === 'left' ? -1 : 1
-    if (this.onCaretWillBeLeaving(path, range, direction)) {
+    if (this.isCaretShouldLeavePath(path, range, direction)) {
       // 跨path移动 先执行跨ptah动作 再执行path内移动动作
       res = this.onCaretLeave(path, range, direction)
     } else {
@@ -238,7 +277,7 @@ export default class Content extends Component {
    * @param {event} [event=null]
    * @memberof Content
    */
-  onLinefeed(range, event = null) {
+  onLinefeed (range, event = null) {
     if (range.inputState.isComposing) return
     event?.preventDefault?.()
     if (!range.collapsed) {
@@ -261,7 +300,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownArrowLeft(range, event) {
+  onKeydownArrowLeft (range, event) {
     horizontalMove('left', range, event)
     this.$editor.selection.updateCaret()
   }
@@ -273,7 +312,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownArrowRight(range, event) {
+  onKeydownArrowRight (range, event) {
     horizontalMove('right', range, event)
     this.$editor.selection.updateCaret()
   }
@@ -285,7 +324,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownArrowUp(range, event) {
+  onKeydownArrowUp (range, event) {
     verticalMove('up', range, event)
     this.$editor.selection.updateCaret()
   }
@@ -297,7 +336,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownArrowDown(range, event) {
+  onKeydownArrowDown (range, event) {
     verticalMove('down', range, event)
     this.$editor.selection.updateCaret()
   }
@@ -308,7 +347,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownBackspace(range) {
+  onKeydownBackspace (range) {
     del(range, false)
   }
 
@@ -319,7 +358,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onKeydownEnter(range, event) {
+  onKeydownEnter (range, event) {
     this.onLinefeed(range, event)
   }
 
@@ -330,7 +369,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onInput(range, event) {
+  onInput (range, event) {
     input(range, event)
   }
 
@@ -343,7 +382,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  onCaretWillBeLeaving(path, range, direction) {
+  isCaretShouldLeavePath (path, range, direction) {
     if (direction === 'left' && range.offset <= 1) {
       let toPath = path.prevLeaf
       if (!toPath) return false
@@ -367,7 +406,7 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  _updatePoints(container, position, distance, newContainer) {
+  _updatePoints (container, position, distance, newContainer) {
     this.$editor.selection.updatePoints(container, position, distance, newContainer)
   }
 
@@ -378,11 +417,11 @@ export default class Content extends Component {
    * @memberof Content
    * @instance
    */
-  setFormat(range, callback) {
+  setFormat (range, callback) {
     const commonPath = this.$editor.queryCommonPath(range.startContainer, range.endContainer)
     const selectedPath = this.$editor.selection.getSeletedPath()
     for (const path of selectedPath) {
-      callback(path.node.formats)
+      callback(path)
     }
     commonPath.component.update()
   }
